+++ app/src/Application.cpp
@@ -48,20 +48,7 @@ void Application::Init(DrawData drawData)
 	Window::WindowController::Init(vulkanBackend);
 	Window::WindowController::GetInstance().NewWindow();
 
-	Input::InputManager::Init();
-
-
-	// Setup Dear ImGui context
-	IMGUI_CHECKVERSION();
-	ImGui::CreateContext();
-	ImGuiIO &io = ImGui::GetIO();
-	(void)io;
-	// io.ConfigFlags |= ImGuiConfigFlags_NavEnableKeyboard;     // Enable Keyboard Controls
-	// io.ConfigFlags |= ImGuiConfigFlags_NavEnableGamepad;      // Enable Gamepad Controls
-
-	// Setup Dear ImGui style
-	ImGui::StyleColorsDark();
-	// ImGui::StyleColorsClassic();
+	Core::Input::InputManager::Init();
 
 	GLFWwindow *glfwWindow = Window::WindowController::GetInstance().GetWindow();
 	// ImGui_ImplGlfw_InitForVulkan(glfwWindow, false);
@@ -100,17 +87,19 @@ void Application::Init(DrawData drawData)
 
 	vulkanBackend.SetupVulkan(extensions, extensions_count, drawData.vertices, drawData.indices, debugBufferSize, debug);
 
-	glfwSetCursorPosCallback(glfwWindow, Input::cursor_position_callback);
-	glfwSetMouseButtonCallback(glfwWindow, Input::mouse_button_callback);
+	glfwSetCursorPosCallback(glfwWindow, Core::Input::cursor_position_callback);
+	glfwSetMouseButtonCallback(glfwWindow, Core::Input::mouse_button_callback);
 
-	Input::InputManager::GetInstance().AddCallback([this](const Input::MouseEvent &e)
-												   {
-		debug_trace_coord = glm::vec2(e.x, e.y);
-		drawn = false; });
+	// Input::InputManager::GetInstance().AddCallback([this](const Input::MouseEvent &e)
+	// 											   {
+	// 	debug_trace_coord = glm::vec2(e.x, e.y);
+	// 	drawn = false; });
 
 	items[0] = "Pixel Debug";
 	items[1] = "All";
 	items[2] = "None";
+
+	// vulkanBackend.init_imgui();
 
 	// ImGui_ImplVulkanH_Window *wd = &g_MainWindowData;
 	// SetupVulkanWindow(wd, surface, w, h);
@@ -171,7 +160,7 @@ void Application::RenderLoop(double t, double fpsLimit)
 
 	std::vector<Utilities::DebugLine> d_lines;
 
-	Input::InputManager::GetInstance().ProcessEvents();
+	Core::Input::InputManager::GetInstance().ProcessEvents();
 
 	// if (!drawn)
 	//{
@@ -191,7 +180,7 @@ void Application::RenderLoop(double t, double fpsLimit)
 	if (m_FrameIndex == 1)
 		memset(m_AccumulationData, 0, static_cast<size_t>(m_Image->GetWidth()) * static_cast<size_t>(m_Image->GetHeight()) * sizeof(glm::vec4));
 
-	glm::vec2 mouse = Input::InputManager::GetInstance().GetMousePosition();
+	glm::vec2 mouse = Core::Input::InputManager::GetInstance().GetMousePosition();
 
 	// std::thread::hardwareconcurrency();
 
@@ -215,108 +204,6 @@ void Application::RenderLoop(double t, double fpsLimit)
 	}
 
 	glm::vec2 size = Window::WindowController::GetInstance().GetSize();
-	// editor->CalculateLayout(size.x, size.y);
-	/*DrawData drawData = editor->RenderEditor();*/
-
-	// imgui new frame
-	ImGui_ImplVulkan_NewFrame();
-	// ImGui_ImplSDL2_NewFrame();
-	ImGui_ImplGlfw_NewFrame();
-	ImGui::NewFrame();
-
-	//----------------------------------------------------------------------------
-	// ImGui UI Instructions
-	ImGUI::Panel::Begin();
-
-	ImGui::SeparatorText("RENDER SETTINGS");
-	ImGui::DragInt("Samples per Pixel 0..50", &samples_per_pixel, 1, 1, 50, "%d%", ImGuiSliderFlags_AlwaysClamp);
-	ImGui::DragInt("Max Bounces 0..30", &max_bounces, 1, 1, 30, "%d%", ImGuiSliderFlags_AlwaysClamp);
-	ImGui::Checkbox("Accumulate", &accumulate);
-
-	ImGui::SeparatorText("DEBUG SETTINGS");
-	bool changed = false;
-
-	ImGui::Combo("Debug Mode", &item_current, items, IM_ARRAYSIZE(items));
-
-	ImGui::SeparatorText("SCENE SETTINGS");
-
-	ImGui::Text("Objects");
-	for (size_t i = 0; i < scene.ray_targets.size(); ++i)
-	{
-		ImGui::PushID(i);
-		std::shared_ptr<Scene::Shapes::RayTarget> val = scene.ray_targets[i];
-		auto sphere = std::dynamic_pointer_cast<Scene::Shapes::Sphere>(val);
-		if (sphere)
-		{
-			ImGui::Text("Sphere \n");
-			float *_origin_ptr = &sphere.get()->Origin.x;
-			ImGui::DragFloat3("Origin", _origin_ptr, 0.1f);
-			ImGui::DragFloat("Radius", &sphere.get()->Radius, 1.0f, 0.01f, 100.0f, "%.2f", ImGuiSliderFlags_AlwaysClamp);
-
-			if (ImGui::BeginCombo("Material", sphere.get()->Material ? sphere.get()->Material->GetName().c_str() : "None"))
-			{
-				for (size_t i = 0; i < scene.materials.size(); ++i)
-				{
-					bool isSelected = (scene.materials[i] == sphere.get()->Material);
-					if (ImGui::Selectable(scene.materials[i]->GetName().c_str(), isSelected))
-					{
-						sphere.get()->Material = scene.materials[i];
-					}
-					if (isSelected)
-						ImGui::SetItemDefaultFocus();
-				}
-				ImGui::EndCombo();
-			}
-		}
-		ImGui::PopID();
-	}
-
-	ImGui::Text("Materials");
-
-	for (size_t i = 0; i < scene.materials.size(); i++)
-	{
-		ImGui::PushID(i);
-
-		std::shared_ptr<Materials::Material> material = scene.materials[i];
-
-		ImGui::Text(material.get()->GetName().c_str());
-
-		float *color_ptr = &material.get()->Albedo.x;
-		ImGui::ColorEdit3("Color", color_ptr);
-
-		auto emissive = std::dynamic_pointer_cast<Materials::Emissive>(material);
-
-		if (emissive)
-		{
-			ImGui::DragFloat("Emission Strength", &emissive.get()->EmissionStrength, 0.05f, 0.0f, 1.0f);
-			float *emission_color_ptr = &emissive.get()->EmissionColor.x;
-			ImGui::ColorEdit3("Emission Color", emission_color_ptr);
-		}
-		else
-		{
-
-			auto metal = std::dynamic_pointer_cast<Materials::Metal>(material);
-			if (metal)
-				ImGui::DragFloat("Fuzz Amt", &metal.get()->fuzz, 0.05f, 0.0f, 1.0f);
-		}
-
-		// ImGui::DragFloat("Roughness", &material.Roughness, 0.05f, 0.0f, 1.0f);
-		// ImGui::DragFloat("Metallic", &material.Metallic, 0.05f, 0.0f, 1.0f);
-
-		ImGui::Separator();
-		ImGui::PopID();
-	}
-
-	ImGui::Text("Last Render: %.3fms", m_LastRenderTime);
-
-	changed |= ImGui::Checkbox("Trace rays", &useRaytracer);
-
-	if (ImGui::Button("Save scene"))
-		sceneManager.SaveScene(scene);
-
-	ImGUI::Panel::End();
-
-	//----------------------------------------------------------------------------
 
 	vulkanBackend.drawFrame(m_DrawData.indices, d_lines.size());
