+++ app/src/Application.hpp
@@ -14,9 +14,9 @@
 #include "InputManager.hpp"
 // #include "../../editor/editor.hpp"
 
-#include "imgui.h"
-#include "imgui_impl_glfw.h"
-#include "imgui_impl_vulkan.h"
+//#include "imgui.h"
+//#include "imgui_impl_glfw.h"
+//#include "imgui_impl_vulkan.h"
 #include "SceneManager.hpp"
 
 struct DrawData
