+++ app/src/VulkanBackend.cpp
@@ -37,87 +37,6 @@ void VulkanBackend::SetupVulkan(const char **extensions, uint32_t extensions_cou
 
     maxDebugLines = 1572864;
     debug = _debug;
-
-    init_imgui();
-}
-
-void VulkanBackend::init_imgui()
-{
-    // 1: create descriptor pool for IMGUI
-    //  the size of the pool is very oversize, but it's copied from imgui demo
-    //  itself.
-    VkDescriptorPoolSize pool_sizes[] = {{VK_DESCRIPTOR_TYPE_SAMPLER, 1000},
-                                         {VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, 1000},
-                                         {VK_DESCRIPTOR_TYPE_SAMPLED_IMAGE, 1000},
-                                         {VK_DESCRIPTOR_TYPE_STORAGE_IMAGE, 1000},
-                                         {VK_DESCRIPTOR_TYPE_UNIFORM_TEXEL_BUFFER, 1000},
-                                         {VK_DESCRIPTOR_TYPE_STORAGE_TEXEL_BUFFER, 1000},
-                                         {VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER, 1000},
-                                         {VK_DESCRIPTOR_TYPE_STORAGE_BUFFER, 1000},
-                                         {VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC, 1000},
-                                         {VK_DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC, 1000},
-                                         {VK_DESCRIPTOR_TYPE_INPUT_ATTACHMENT, 1000}};
-
-    VkDescriptorPoolCreateInfo pool_info = {};
-    pool_info.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO;
-    pool_info.flags = VK_DESCRIPTOR_POOL_CREATE_FREE_DESCRIPTOR_SET_BIT;
-    pool_info.maxSets = 1000;
-    pool_info.poolSizeCount = (uint32_t)std::size(pool_sizes);
-    pool_info.pPoolSizes = pool_sizes;
-
-    VkDescriptorPool imguiPool;
-    VkResult err = vkCreateDescriptorPool(g_Device, &pool_info, nullptr, &imguiPool);
-    check_vk_result(err);
-
-    // 2: initialize imgui library
-
-    // this initializes the core structures of imgui
-    ImGui::CreateContext();
-    ImGUI::setStyle(true);
-
-    ImGuiIO &io = ImGui::GetIO();
-    (void)io;
-    io.ConfigFlags |= ImGuiConfigFlags_NavEnableKeyboard; // Enable Keyboard Controls
-    io.ConfigFlags |= ImGuiConfigFlags_NavEnableGamepad;  // Enable Gamepad Controls
-    io.ConfigFlags |= ImGuiConfigFlags_DockingEnable;
-
-    // Scale style sizes for high-DPI monitors
-    ImGuiStyle &imgui_style = ImGui::GetStyle();
-    imgui_style.ScaleAllSizes(ImGUI::getDPIScale());
-
-    // this initializes imgui for SDL
-    // ImGui_ImplSDL2_InitForVulkan(_window);
-
-    ImGui_ImplGlfw_InitForVulkan(Window::WindowController::GetInstance().GetWindow(), false);
-
-    // this initializes imgui for Vulkan
-    ImGui_ImplVulkan_InitInfo init_info = {};
-    init_info.Instance = g_Instance;
-    init_info.PhysicalDevice = g_PhysicalDevice;
-    init_info.Device = g_Device;
-    init_info.Queue = g_GraphicsQueue;
-    init_info.DescriptorPool = imguiPool;
-    init_info.MinImageCount = 3;
-    init_info.ImageCount = 3;
-    init_info.RenderPass = context.RenderPass;
-
-    // init_info.UseDynamicRendering = true;
-
-    // VkPipelineRenderingCreateInfoKHR vkPipelineREnderingCreateInfoKHR = {};
-    // vkPipelineREnderingCreateInfoKHR.sType = VK_STRUCTURE_TYPE_PIPELINE_RENDERING_CREATE_INFO;
-    // vkPipelineREnderingCreateInfoKHR.colorAttachmentCount = 1;
-    // vkPipelineREnderingCreateInfoKHR.pColorAttachmentFormats = &swapChainImageFormat;
-
-    //// dynamic rendering parameters for imgui to use
-    // init_info.PipelineRenderingCreateInfo = vkPipelineREnderingCreateInfoKHR;
-    // init_info.MSAASamples = VK_SAMPLE_COUNT_1_BIT;
-
-    ImGui_ImplVulkan_Init(&init_info);
-
-    ImGui_ImplVulkan_CreateFontsTexture();
-
-    vkDeviceWaitIdle(g_Device);
-    // ImGui_ImplVulkan_DestroyFontUploadObjects();
 }
 
 void VulkanBackend::drawFrame(const std::vector<uint16_t> indices, int debugLinesCount)
@@ -785,7 +704,9 @@ void VulkanBackend::recordCommandBuffer(VkCommandBuffer commandBuffer, const std
     if (debug)
         drawDebugRays(commandBuffer, debugLinesCount);
 
-    draw_imgui(commandBuffer);
+    //draw_imgui(commandBuffer);
+
+    if (recordCallback) recordCallback(commandBuffer);
 
     vkCmdEndRenderPass(commandBuffer);
 
@@ -1264,10 +1185,4 @@ void VulkanBackend::drawDebugRays(VkCommandBuffer commandBuffer, int numLines)
     vkCmdBindPipeline(commandBuffer, VK_PIPELINE_BIND_POINT_GRAPHICS, context.DebugPipeline);
     vkCmdBindVertexBuffers(commandBuffer, 0, 1, &debugVertexBuffer, offsets);
     vkCmdDraw(commandBuffer, numLines * 2, 1, 0, 0);
-}
-
-void VulkanBackend::draw_imgui(VkCommandBuffer cmd)
-{
-    ImGui::Render();
-    ImGui_ImplVulkan_RenderDrawData(ImGui::GetDrawData(), cmd);
 }
